+++ src/abe_policy/interfaces/mod.rs
@@ -1,5 +1,0 @@
-#[cfg(feature = "wasm_bindgen")]
-pub mod wasm_bindgen;
-
-#[cfg(feature = "ffi")]
-pub mod ffi;