+++ src/interfaces/ffi/mod.rs
@@ -1,5 +1,0 @@
-pub mod generate_cc_keys;
-pub mod hybrid_cc_aes;
-
-#[cfg(test)]
-mod tests;