+++ CHANGELOG.md
@@ -4,6 +4,33 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+## [11.0.0] - 2023-03-01
+
+### Bug Fixes
+
+- Decrypt test example
+
+### Documentation
+
+- Update readme
+
+### Refactor
+
+- [**breaking**] Move all interfaces (FFI, Wasm, pyo3) to `cloudproof_rust` repository
+- remove unneeded dependencies
+- remove inline macros
+- remove serde wherever possible
+
+### Ci
+
+- Add autopublish action
+
+### Testing
+
+- Re-expose non-regression mechanism
+
+---
+
 ## [10.0.0] - 2023-02-02
 
 ### Documentation
@@ -19,6 +46,8 @@ All notable changes to this project will be documented in this file.
 
 - Update KMS version
 - Rebase externals repos
+
+---
 
 ## [9.0.0] - 2023-01-20